dark mode support: use colors from style
Instead of hardcoding colors in code - we now use colors from style.
This makes dark mode work a little better, and the UI looks more ...
boring and native.

Some colors have an alpha channel, to make them less "in your face",
which seems to look nice on Plasma5 on light and dark themes.