Initial code for theme support
Basically I can load a theme. I can also apply (some parts) of the theme
into the QPlainTextEdit (QutePart). Then - when loading a syntax, I
"can" apply the theme when loading the styles.

So far, I am not having any success in this. The contexts I find while
parsing the syntax do not match the ones on the style.

``` xml
    <list name="controlflow">
        <keyword attribute="Control Flow" context="#stay"
String="controlflow" />
      <itemData name="Control Flow" defStyleNum="dsControlFlow"
spellChecking="false"/>

```

It seems I am matching against the `attribute` and not the `String`.
Working on that, then I will clean up the code.