Theme: allow to chage a live syntax's theme
When we tried to load a lanague, with  new theme (effetively change the
colors cheme of the language) this would originall fail, as the code by
default contains a single definition of a language (which is a sane
idea).

The syntax should be part of the highlighter and not language. This will reduce some
memory consumption, as many files can share the same lanaguge. Otherwise - we cannot
load the same syntax with two different highlighters, nor  change the theme
of languages.