Fix smart end cursor movement
Smart end was always going to the penultimate position. Actually the
cursor was at the left of the last character, instead of the right.

The solution was to put the cursor at "end+1". Why was not it done
before? since sometimes end+1 was out of bounds. This was easy to fix.

closes #17