Qutepart: display white space at ends of lines
Login was too complex for me to understand. I actually threw it to
perplexity and it fixed it. IMHO - this logic is mjust bunkers. To be
revisited in the future.

For now - whitepsace is peoperly displayed at the end of lines. Tabs now
follow the same color scheme as spaces.

Fixes #7