Minimap: show/hide was not working
On Fedora/gnome (at least, I think it worked on Debian/Plasma6) when I
hide and then re-show the minimap it would not be visible.

The solution was to make the minimap visible by calling `show()`. I took
the time to break from the pattern of this library of using C++ shared
pointers to raw pointers, as Qt code does.