All: shave off another 10-15 msec
- While iterating formats, use a reference, not copy.
- While matching, use a naked pointer, instead of scoped pointer.

The last point seems to be the one which brings the test to 2393msec
loading time (before the best I saw was 2408msec, Started with
2466msec).

Testing is done several times, to easy OS caching and different
processes runnning. Taking the lower run as the benchmark.

Shave ~100 msec, with very simple tricks. I might be able to make more.
Need to find good places to fix.