Language: fix bookmarks saving
Bookmarks are saved as a bit, in the block state.
However - the highlighing ode unneededly modified that state as well.
The highlighing code uses an different mechanism to determinate the
block state (`Language::getContextStack()` which internally uses
`userBlock()`).

Which means, all the dynamic cast, can be modified to static_cast,
reducing runtime.