Language: fix bookmarks - #26
The bookmark feature was implemente using `QTextBlock::setUserStatus()`,
this would set a bit and makr that block as "bookmark".

However - as part of the highlighing, that same data was set to the
current context. Why? Because when doing syntax highlighters, you can
maintain the state of your parsing using that integer, or using user
data on a block.

This project uses a custom data block anyway - so the status number was
a mistake.

Now bookmakrs are persistant.

Also - while we are at that:

1) We know everywhere that the block data is ours, we can `static_cast`
it instead of `dynamic_cast`.
2) Remove all `dynamic_cast` from code, and port it to `static_cast`.

Loading our benchmark file:
before: File  "/home/diego/benchmark.cpp" loaded in 2750 mSec
after: File  "/home/diego/benchmark.cpp" loaded in 2684 mSec

closes #26