Qutepart: add support for line markings - 25
You can mark a line as having a warning, error, info, currently executed
breakpoint etc. All these lines will have a different background. The
line background is also seen in the minimap.

Lines can also have a messagea attached. This can be used for displaying
CVS info, or build errors. The text is shows a a tooltip on the side
panels.

closes #25