Completer: fix a crash
When deleting the object, inside the click event, this triggers a nasty
crash. The code fails inside:

```
QAbstractItemView::mouseReleaseEvent(QMouseEvent *event)
```

While accesing the `d` pointer. This event is called right after the
call to the slot that is connected to the function that deletes the
widget. Seems trivial, meaning this code was not executed up until now.

Solution is simple: instead of `delete listview` we can call `listview-
>deleteLater()`. This way the owner will delete the widget, after all
events have been called.