Qutepart: better comment support (#9)
Each context has now another link to the main language. Meaning, when we
toggle comments - we can query the language directly, and not the
document.

Why does it matter?

Imagine a PHP file. Depending on the line, we can be in:

1) php syntax.
2) HTML syntax.
3) CSS synaxt.
4) Javascript syntax.

In such mixed context file, I am able to comment text in all but JS
context. For some reason, the context I am getting is still HTML.

refs #9