Language: use set user state properly (#41)
In order for the syntax highligheter to properly work, we need to set
the user state of the block. However, this was also used for setting the
user flags. Which means both features were using the same API.

If we move the flags to the user block, and inside the highlightrer use
the userstate (as the code used before...) code now works as expected.

closes #41